fix rules
