+++ tests/conftest.py
@@ -40,48 +40,7 @@ def log_stream():
     handler.close()
 
 
-@pytest.fixture
-def rule_processor(caller_options):
-    return RuleProcessor(caller_options)
 
-
-@pytest.fixture
-def caller_options():
-    return {
-        'adobe_group_filter': None,
-        'after_mapping_hook': None,
-        'default_country_code': 'US',
-        'delete_strays': False,
-        'directory_group_filter': None,
-        'disentitle_strays': False,
-        'exclude_groups': [],
-        'exclude_identity_types': ['adobeID'],
-        'exclude_strays': False,
-        'exclude_users': [],
-        'extended_attributes': None,
-        'process_groups': True,
-        'max_adobe_only_users': 200,
-        'new_account_type': 'federatedID',
-        'remove_strays': True,
-        'strategy': 'sync',
-        'stray_list_input_path': None,
-        'stray_list_output_path': None,
-        'test_mode': True,
-        'update_user_info': False,
-        'username_filter_regex': None,
-        'adobe_only_user_action': ['remove'],
-        'adobe_only_user_list': None,
-        'adobe_users': ['all'],
-        'config_filename': 'tests/fixture/user-sync-config.yml',
-        'connector': 'ldap',
-        'encoding_name': 'utf8',
-        'user_filter': None,
-        'users': None,
-        'directory_connector_type': 'csv',
-        'directory_connector_overridden_options': {
-            'file_path': '../tests/fixture/remove-data.csv'},
-        'adobe_group_mapped': False,
-        'additional_groups': []}
 
 @pytest.fixture
 def mock_directory_user():
